Fix lint

+++ solutions/projecteuler/015/015.py
@@ -2,4 +2,4 @@ from math import factorial
 
 n = 20
 
-print(factorial(2 * n) / (factorial(n) ** 2))
+print(int(factorial(2 * n) / (factorial(n) ** 2)))
